+++ src/jaudio/filter3d.c
@@ -6,9 +6,9 @@
  * Address:	8001BB60
  * Size:	000034
  */
-float V3D_GetAngle(struct Vector3D_* input_vec)
+f32 V3D_GetAngle(struct Vector3D_* input_vec)
 {
-	float retval = (atanf2(input_vec->x, input_vec->y) + 3.141592);
+	f32 retval = (atanf2(input_vec->x, input_vec->y) + 3.141592);
 	return retval;
 }
 /*
@@ -36,7 +36,7 @@ void V3D_Minus(Vector3D_*, Vector3D_*, Vector3D_*)
  * Address:	8001BBA0
  * Size:	000040
  */
-double V3D_Abs(struct Vector3D_* input_vec)
+f64 V3D_Abs(struct Vector3D_* input_vec)
 {
 	return (sqrtf2((input_vec->x * input_vec->x) + (input_vec->y * input_vec->y) + (input_vec->z * input_vec->z)));
 }
@@ -56,7 +56,7 @@ void V3D_Norm(Vector3D_*)
  * Address:	........
  * Size:	00005C
  */
-void V3D_Resize(Vector3D_*, float)
+void V3D_Resize(Vector3D_*, f32)
 {
 	// UNUSED FUNCTION
 }
@@ -96,7 +96,7 @@ void V3D_UpdateObj(Object3D_*, Vector3D_*, Vector3D_*)
  * Address:	........
  * Size:	000040
  */
-void V3D_TMove(TVector3D_*, float)
+void V3D_TMove(TVector3D_*, f32)
 {
 	// UNUSED FUNCTION
 }
@@ -116,7 +116,7 @@ void V3D_TimeUpdateObj(Object3D_*)
  * Address:	........
  * Size:	000038
  */
-void V3D_Filter_Set(Filter_*, float*, unsigned long)
+void V3D_Filter_Set(Filter_*, f32*, u32)
 {
 	// UNUSED FUNCTION
 }
@@ -136,7 +136,7 @@ void V3D_Filter_Mix(Filter_*, Filter_*, Filter_*)
  * Address:	........
  * Size:	0000D8
  */
-void V3D_Filter_MixD(Filter_*, float*, float*, unsigned long, unsigned long)
+void V3D_Filter_MixD(Filter_*, f32*, f32*, u32, u32)
 {
 	// UNUSED FUNCTION
 }
